// src/types/speciality.ts
export type Speciality = {
  id: 1;
  name_en: "Dentist";
  name_ar: "طبيب أسنان";
  description: "Dental and oral care specialist";
  icon: "specialists_images/dentist-icon.svg";
  status: 1;
  created_at: "2025-09-16T19:44:08.000000Z";
  updated_at: "2025-09-16T19:44:08.000000Z";
};